apply sin/cos optimization to all decoders

+++ nebula_decoders/include/nebula_decoders/nebula_decoders_hesai/decoders/pandar_64.hpp
@@ -57,7 +57,7 @@ struct Header
 
 struct Unit
 {
-  double distance;
+  float distance;
   uint16_t intensity;
 };
